pyproject.toml
[tool.poetry]
name = "auth-sprint-1"
version = "0.1.0"
description = ""
authors = ["Schislyaev <pschhhh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^1.4.45"
flask = "^2.2.2"
redis = "^4.4.2"
flask-security = "^3.0.0"
flask-jwt-extended = "^4.4.4"
flask-marshmallow = "^0.14.0"
flasgger = "^0.9.5"
alembic = "^1.9.1"
psycopg2-binary = "^2.9.5"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
python-dotenv = "^0.21.0"
flask-sqlalchemy = "^3.0.2"
bcrypt = "^4.0.1"
email-validator = "^1.3.0"
passlib = "^1.7.4"
aiohttp = "^3.8.3"
anyio = "^3.6.2"
pytest-asyncio = "^0.20.3"
pytest = "^7.2.0"
validators = "^0.20.0"
flask-apispec = "^0.11.4"
requests = "^2.28.1"
status = "^0.2.5"
flask-api = "^3.0.post1"
grpcio = "^1.51.1"
grpcio-tools = "^1.51.1"
gunicorn = "^20.1.0"
opentelemetry-api = "^1.15.0"
opentelemetry-sdk = "^1.15.0"
opentelemetry-instrumentation-flask = "^0.36b0"
opentelemetry-exporter-jaeger = "^1.15.0"
jaeger-client = "^4.8.0"
flask-limiter = {extras = ["redis"], version = "^3.1.0"}
coverage = "^7.0.5"
flask-oauthlib = "^0.9.6"
rauth = "^0.7.3"
google-api-python-client = "^2.73.0"
google-auth-oauthlib = "^0.8.0"

[tool.poetry.group.dev.dependencies]
flake8 = "5.0.4"
isort = "5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
